Add files via upload
My scratch program is a small game that consists of making a dog bounce as many times as possible on a trampoline that is controlled by the left and right arrow keys. To start the game, click on the green flag, and then click on the "start" button on the screen.  The player's score is displayed on the top-left corner of the screen.